+++ allennlp/training/grad_scalers/fairscale_sharded_grad_scaler.py
@@ -1,0 +1,9 @@
+from fairscale.optim.grad_scaler import ShardedGradScaler as _FairScaleShardedGradScaler
+from allennlp.training.grad_scalers.grad_scaler import GradScaler
+
+
+@GradScaler.register("fairscale")
+class FairScaleShardedGradScaler(_FairScaleShardedGradScaler, GradScaler):
+    """
+    FairScale's `ShardedGradScaler`.
+    """