+++ allennlp/training/trainer.py
@@ -25,6 +25,7 @@ from allennlp.nn.parallel.ddp_wrappers import DdpWrapper, TorchDdpWrapper
 from allennlp.training import util as training_util
 from allennlp.training.callbacks import TrainerCallback, SanityChecksCallback, ConsoleLoggerCallback
 from allennlp.training.checkpointer import Checkpointer
+from allennlp.training.grad_scalers import GradScaler
 from allennlp.training.learning_rate_schedulers import LearningRateScheduler
 from allennlp.training.metric_tracker import MetricTracker
 from allennlp.training.momentum_schedulers import MomentumScheduler
@@ -262,6 +263,16 @@ class GradientDescentTrainer(Trainer):
         A `DdpWrapper` to use if running in the distributed setting. If left unspecified,
         the default `TorchDdpWrapper` is used.
 
+        !!! Note
+            You generally shouldn't override this unless you really know what you're doing!
+
+    grad_scaler : `Optional[GradScaler]`, optional (default = `None`)
+        The `GradScaler` to use when training with AMP enabled. If left unspecified, the
+        default PyTorch `GradScaler` will be used.
+
+        !!! Note
+            You generally shouldn't override this unless you really know what you're doing!
+
     """
 
     def __init__(
@@ -290,6 +301,7 @@ class GradientDescentTrainer(Trainer):
         enable_default_callbacks: bool = True,
         run_sanity_checks: bool = True,
         ddp_wrapper: Optional[DdpWrapper] = None,
+        grad_scaler: Optional[GradScaler] = None,
     ) -> None:
         super().__init__(serialization_dir, cuda_device, distributed, local_rank, world_size)
 
@@ -348,12 +360,12 @@ class GradientDescentTrainer(Trainer):
         self._num_gradient_accumulation_steps = num_gradient_accumulation_steps
 
         # Enable automatic mixed precision training.
-        self._scaler: Optional[amp.GradScaler] = None
+        self._scaler: Optional[GradScaler] = None
         self._use_amp = use_amp
         if self._use_amp:
             if self.cuda_device == torch.device("cpu"):
                 raise ValueError("Using AMP requires a cuda device")
-            self._scaler = amp.GradScaler()
+            self._scaler = grad_scaler or GradScaler()
 
         # Using `DistributedDataParallel`(ddp) brings in a quirk wrt AllenNLP's `Model` interface and its
         # usage. A `Model` object is wrapped by `ddp`, but assigning the wrapped model to `self.model`